fix: merge browse-listings into get-listings, center map on SLP (#38)
