+++ packages/infra/src/lambdas/get-listings.mjs
@@ -6,22 +6,24 @@ import { createLogger } from "./logger.mjs";
 const client = new DynamoDBClient({});
 const ddb = DynamoDBDocumentClient.from(client);
 const TABLE = process.env.LISTINGS_TABLE;
+const STATUS_INDEX = process.env.STATUS_INDEX;
 
 export const handler = async (event) => {
   const log = createLogger(event);
   try {
     const userId = getUserId(event);
-    if (!userId) {
-      return {
-        statusCode: 401,
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ error: "Unauthorized" }),
-      };
-    }
-
     const mine = event.queryStringParameters?.mine;
 
+    // mine=true requires authentication
     if (mine === "true") {
+      if (!userId) {
+        return {
+          statusCode: 401,
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({ error: "Unauthorized" }),
+        };
+      }
+
       const result = await ddb.send(
         new QueryCommand({
           TableName: TABLE,
@@ -39,10 +41,51 @@ export const handler = async (event) => {
       };
     }
 
+    // Browse available listings (public or authenticated)
+    const category = event.queryStringParameters?.category;
+    const cursor = event.queryStringParameters?.cursor;
+    const PAGE_SIZE = 20;
+
+    const queryParams = {
+      TableName: TABLE,
+      IndexName: STATUS_INDEX,
+      KeyConditionExpression: "#status = :status",
+      ExpressionAttributeNames: { "#status": "status" },
+      ExpressionAttributeValues: { ":status": "available" },
+      ScanIndexForward: false,
+      Limit: PAGE_SIZE,
+    };
+
+    if (category) {
+      queryParams.FilterExpression = "category = :cat";
+      queryParams.ExpressionAttributeValues[":cat"] = category;
+    }
+
+    if (cursor) {
+      try {
+        queryParams.ExclusiveStartKey = JSON.parse(Buffer.from(cursor, "base64url").toString());
+      } catch {
+        // Invalid cursor, ignore
+      }
+    }
+
+    const result = await ddb.send(new QueryCommand(queryParams));
+
+    const listings = (result.Items || [])
+      .filter((item) => !userId || item.userId !== userId)
+      .map(({ userId: _ownerId, ...item }) => ({
+        ...item,
+        address: redactAddress(item.address),
+      }));
+
+    const nextCursor = result.LastEvaluatedKey
+      ? Buffer.from(JSON.stringify(result.LastEvaluatedKey)).toString("base64url")
+      : null;
+
     return {
-      statusCode: 400,
+      statusCode: 200,
       headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ error: "Query parameter mine=true is required" }),
+      body: JSON.stringify({ listings, nextCursor }),
     };
   } catch (err) {
     log.error("get-listings failed", err);
@@ -53,3 +96,16 @@ export const handler = async (event) => {
     };
   }
 };
+
+/** Strip street address, keep only city/state. */
+function redactAddress(address) {
+  if (!address) return "";
+  const parts = address.split(",").map((p) => p.trim());
+  if (parts.length >= 3) {
+    return `${parts[parts.length - 2]}, ${parts[parts.length - 1].replace(/\d{5}(-\d{4})?/, "").trim()}`;
+  }
+  if (parts.length === 2) {
+    return parts[1].replace(/\d{5}(-\d{4})?/, "").trim();
+  }
+  return "Twin Cities area";
+}
