+++ template.yaml
@@ -62,34 +62,12 @@ Resources:
       Environment:
         Variables:
           LISTINGS_TABLE: !Sub "scrappr-listings-${Env}"
-      Events:
-        Api:
-          Type: HttpApi
-          Properties:
-            Path: /listings
-            Method: GET
-            ApiId: !Ref HttpApi
-
-  BrowseListingsFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      Handler: browse-listings.handler
-      CodeUri: packages/infra/src/lambdas/
-      Environment:
-        Variables:
-          LISTINGS_TABLE: !Sub "scrappr-listings-${Env}"
           STATUS_INDEX: status-index
       Events:
         Api:
           Type: HttpApi
           Properties:
-            Path: /listings/available
-            Method: GET
-            ApiId: !Ref HttpApi
-        PublicApi:
-          Type: HttpApi
-          Properties:
-            Path: /listings/browse
+            Path: /listings
             Method: GET
             ApiId: !Ref HttpApi
